chore: reorganize connection status dir
