+++ services/offline/src/lib/dhis2-connection-status/smart-interval.ts
