+++ services/offline/src/lib/dhis2-connection-status/use-ping-query.ts
